Shard Python Roles tests
Separates into three files in a new
package at test/python/roles. Divided
by testing standard roles requests,
member and membership requests, and
requests including query params.
